+++ ledger/ledger-api-common/src/main/scala/com/digitalasset/ledger/api/validation/TransactionServiceRequestValidator.scala
@@ -4,7 +4,7 @@
 package com.digitalasset.ledger.api.validation
 
 import brave.propagation.TraceContext
-import com.digitalasset.daml.lf.data.Ref.Party
+import com.digitalasset.daml.lf.data.Ref
 import com.digitalasset.ledger.api.domain
 import com.digitalasset.ledger.api.domain.{LedgerId, LedgerOffset}
 import com.digitalasset.ledger.api.messages.transaction
@@ -35,6 +35,8 @@ class TransactionServiceRequestValidator(
 
   private val filterValidator = new TransactionFilterValidator(identifierResolver)
 
+  private val partyValidator = new PartyValidator(partyNameChecker)
+
   private def matchId(input: LedgerId): Result[LedgerId] = matchLedgerId(ledgerId)(input)
 
   private val rightNone = Right(None)
@@ -44,6 +46,7 @@ class TransactionServiceRequestValidator(
       transactionFilter: TransactionFilter,
       begin: domain.LedgerOffset,
       end: Option[domain.LedgerOffset],
+      knownParties: Set[Ref.Party],
       traceContext: Option[TraceContext])
 
   private def commonValidations(req: GetTransactionsRequest): Result[PartialValidation] = {
@@ -55,13 +58,14 @@ class TransactionServiceRequestValidator(
       convertedEnd <- req.end
         .fold[Result[Option[domain.LedgerOffset]]](rightNone)(end =>
           LedgerOffsetValidator.validate(end, "end").map(Some(_)))
-      _ <- requireKnownParties(req.getFilter)
+      knownParties <- partyValidator.requireKnownParties(req.getFilter.filtersByParty.keySet)
     } yield
       PartialValidation(
         ledgerId,
         filter,
         convertedBegin,
         convertedEnd,
+        knownParties,
         req.traceContext.map(toBrave))
 
   }
@@ -79,25 +83,6 @@ class TransactionServiceRequestValidator(
       case _ => Right(())
     }
 
-  private def requireParties(parties: Traversable[String]): Result[Set[Party]] =
-    parties.foldLeft[Result[Set[Party]]](Right(Set.empty)) { (acc, partyTxt) =>
-      for {
-        parties <- acc
-        party <- requireParty(partyTxt)
-      } yield parties + party
-    }
-
-  //TODO PartyValidator
-  private def requireKnownParties(transactionFilter: TransactionFilter): Result[Unit] =
-    requireParties(transactionFilter.filtersByParty.keys).flatMap(requireKnownParties)
-
-  private def requireKnownParties(partiesInRequest: Iterable[Party]): Result[Unit] = {
-    val unknownParties = partiesInRequest.filterNot(partyNameChecker.isKnownParty)
-    if (unknownParties.nonEmpty)
-      Left(invalidArgument(s"Unknown parties: ${unknownParties.mkString("[", ", ", "]")}"))
-    else Right(())
-  }
-
   def validate(
       req: GetTransactionsRequest,
       ledgerEnd: LedgerOffset.Absolute,
@@ -112,7 +97,6 @@ class TransactionServiceRequestValidator(
       convertedFilter <- filterValidator.validate(
         partial.transactionFilter,
         "filter.filters_by_party")
-      _ <- requireKnownParties(req.getFilter)
     } yield {
       transaction.GetTransactionsRequest(
         ledgerId,
@@ -163,8 +147,7 @@ class TransactionServiceRequestValidator(
       _ <- requireNumber(req.transactionId, "transaction_id")
       trId <- requireLedgerString(req.transactionId)
       _ <- requireNonEmpty(req.requestingParties, "requesting_parties")
-      parties <- requireParties(req.requestingParties)
-      _ <- requireKnownParties(parties)
+      parties <- partyValidator.requireKnownParties(req.requestingParties)
     } yield {
       transaction.GetTransactionByIdRequest(
         ledgerId,
@@ -180,8 +163,7 @@ class TransactionServiceRequestValidator(
       ledgerId <- matchId(LedgerId(req.ledgerId))
       eventId <- requireLedgerString(req.eventId, "event_id")
       _ <- requireNonEmpty(req.requestingParties, "requesting_parties")
-      parties <- requireParties(req.requestingParties)
-      _ <- requireKnownParties(parties)
+      parties <- partyValidator.requireKnownParties(req.requestingParties)
     } yield {
       transaction.GetTransactionByEventIdRequest(
         ledgerId,
@@ -201,6 +183,6 @@ class TransactionServiceRequestValidator(
           invalidArgument(
             s"$party attempted subscription for templates ${inclusive.templateIds.mkString("[", ", ", "]")}. Template filtration is not supported on GetTransactionTrees RPC. To get filtered data, use the GetTransactions RPC.")
       }
-      .fold(requireParties(transactionFilter.filtersByParty.keys))(Left(_))
+      .fold(partyValidator.requireKnownParties(transactionFilter.filtersByParty.keys))(Left(_))
 
 }
